Moved all gui files to the root gui folder
