+++ nw/gui/about.py
