--- nw/gui/dialogs/__init__.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from nw.gui.dialogs.about import GuiAbout
-from nw.gui.dialogs.preferences import GuiPreferences
-from nw.gui.dialogs.docmerge import GuiDocMerge
-from nw.gui.dialogs.docsplit import GuiDocSplit
-from nw.gui.dialogs.itemeditor import GuiItemEditor
-from nw.gui.dialogs.projectsettings import GuiProjectSettings
-from nw.gui.dialogs.projectload import GuiProjectLoad
-from nw.gui.dialogs.sessionlog import GuiSessionLogView
-
-__all__ = [
-    "GuiAbout",
-    "GuiPreferences",
-    "GuiDocMerge",
-    "GuiDocSplit",
-    "GuiItemEditor",
-    "GuiProjectSettings",
-    "GuiProjectLoad",
-    "GuiSessionLogView",
-]
