+++ nw/gui/docbars.py
