+++ nw/gui/docdetails.py
