--- nw/gui/elements/__init__.py
@@ -1,23 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from nw.gui.elements.docdetails import GuiDocDetails
-from nw.gui.elements.doceditor import GuiDocEditor
-from nw.gui.elements.docbars import GuiDocTitleBar
-from nw.gui.elements.doctree import GuiDocTree
-from nw.gui.elements.docviewer import GuiDocViewer
-from nw.gui.elements.docbars import GuiNoticeBar
-from nw.gui.elements.outline import GuiProjectOutline
-from nw.gui.elements.docbars import GuiSearchBar
-from nw.gui.elements.viewdetails import GuiDocViewDetails
-
-__all__ = [
-    "GuiDocDetails",
-    "GuiDocEditor",
-    "GuiDocTitleBar",
-    "GuiDocTree",
-    "GuiDocViewer",
-    "GuiNoticeBar",
-    "GuiProjectOutline",
-    "GuiSearchBar",
-    "GuiDocViewDetails",
-]
