+++ nw/gui/itemeditor.py
