+++ nw/gui/outline.py
