+++ nw/gui/projectsettings.py
