+++ nw/gui/sessionlog.py
