+++ nw/gui/viewdetails.py
